% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp3tools.R
\name{classify_gazepoint_export}
\alias{classify_gazepoint_export}
\title{Classify a Gazepoint export}
\usage{
classify_gazepoint_export(path)
}
\arguments{
\item{path}{File path.}
}
\value{
A single character string.
}
\description{
Uses the filename and header structure to classify a file as \code{all_gaze},
\code{fixations}, \code{summary}, or \code{unknown}.
}
