% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil_signal_extensions.R
\name{mean_gazepoint_pupil}
\alias{mean_gazepoint_pupil}
\title{Calculate mean binocular pupil size}
\usage{
mean_gazepoint_pupil(
  master_df,
  lp_col = "LPupil",
  rp_col = "RPupil",
  output_col = "mean_pupil",
  min_eyes = 1
)
}
\arguments{
\item{master_df}{A sample-level pupil data frame.}

\item{lp_col, rp_col}{Left- and right-pupil columns.}

\item{output_col}{Name of the generated column.}

\item{min_eyes}{Minimum number of finite eye measurements required.}
}
\value{
The input data with a binocular mean-pupil column.
}
\description{
Calculate mean binocular pupil size
}
\examples{
pupil <- data.frame(
  LPupil = c(3, NA, 4),
  RPupil = c(3.2, 3.5, NA)
)
mean_gazepoint_pupil(pupil)
}
