% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp3tools.R
\name{plot_transition_heatmap}
\alias{plot_transition_heatmap}
\title{Plot an AOI transition heatmap}
\usage{
plot_transition_heatmap(transitions)
}
\arguments{
\item{transitions}{Output of \code{compute_transition_matrix()}.}
}
\value{
A ggplot object.
}
\description{
Plot an AOI transition heatmap
}
