% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp3tools.R
\name{summarise_fixations}
\alias{summarise_fixations}
\title{Summarise fixation-level AOI metrics}
\usage{
summarise_fixations(data, group_cols = "MEDIA_ID", aoi_col = "AOI")
}
\arguments{
\item{data}{A Gazepoint fixation data frame.}

\item{group_cols}{Grouping columns.}

\item{aoi_col}{AOI column name.}
}
\value{
A tibble with fixation counts and summed fixation duration by AOI.
}
\description{
Summarise fixation-level AOI metrics
}
