% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp3tools.R
\name{summarise_tracking_quality}
\alias{summarise_tracking_quality}
\title{Summarise Gazepoint tracking quality}
\usage{
summarise_tracking_quality(data, group_cols = "MEDIA_ID")
}
\arguments{
\item{data}{A Gazepoint data frame.}

\item{group_cols}{Grouping columns.}
}
\value{
A tibble with validity percentages for available validity columns.
}
\description{
Summarise Gazepoint tracking quality
}
