% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_benchmarking.R
\name{write_gazepoint_performance_benchmark}
\alias{write_gazepoint_performance_benchmark}
\title{Write gp3tools performance benchmark tables}
\usage{
write_gazepoint_performance_benchmark(
  x,
  output_dir,
  prefix = "gp3tools-performance"
)
}
\arguments{
\item{x}{A \code{"gazepoint_performance_benchmark"} object.}

\item{output_dir}{Output directory.}

\item{prefix}{Filename prefix.}
}
\value{
Named character vector of written files.
}
\description{
Write gp3tools performance benchmark tables
}
